Final tests and cleanup
Moved sleeptimer to main activity and made it actually work
